See release notes.

+++ nautilus_core/persistence/tests/test_util.rs
@@ -1,5 +1,5 @@
 // -------------------------------------------------------------------------------------------------
-//  Copyright (C) 2015-2022 Nautech Systems Pty Ltd. All rights reserved.
+//  Copyright (C) 2015-2023 Nautech Systems Pty Ltd. All rights reserved.
 //  https://nautechsystems.io
 //
 //  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
@@ -13,10 +13,9 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use rand::Rng;
-
 use nautilus_model::types::fixed::f64_to_fixed_i64;
 use nautilus_model::types::fixed::f64_to_fixed_u64;
+use rand::Rng;
 
 #[allow(dead_code)]
 fn random_values_u64(len: u64) -> Vec<u64> {
@@ -47,7 +46,7 @@ fn random_values_u8(len: u64) -> Vec<u8> {
     let mut rng = rand::thread_rng();
     let mut vec = Vec::new();
     for _ in 0..len {
-        let value = rng.gen_range(2..5) as u8;
+        let value = rng.gen_range(2..5);
         vec.push(value);
     }
     assert_eq!(vec.len() as u64, len);
